% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{add_age_mid}
\alias{add_age_mid}
\title{Add a column called '\if{html}{\out{<agevar>}}.mid' to df}
\usage{
add_age_mid(df, agevar, age_width_df, age_min)
}
\arguments{
\item{df}{Data frame containing an age variable}

\item{agevar}{Name of the age variable}

\item{age_width_df}{Data frame with columns "age" and "width"}

\item{age_min}{The lower limit of the youngest age group}
}
\value{
'df' with an extra column called 'age.mid'
}
\description{
Column '\if{html}{\out{<agevar>}}.mid' gives the midpoint of
each age group.
}
