% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-constructors.R
\name{AR1}
\alias{AR1}
\title{Specify an AR1 model for a time effect}
\usage{
AR1(coef_min = 0.8, coef_max = 0.98, mean = 4, sd = 2, scale = 1)
}
\arguments{
\item{coef_min, coef_max}{Lower and upper limits
for the autocorrelation coefficient.
Defaults are 0.80 and 0.98.}

\item{mean, sd}{Mean and standard
deviation for the prior for
the long-run average (on the log scale).
Defaults are 4 and 2.}

\item{scale}{Scale for
standard deviation
of error term. Defaults to 1.}
}
\value{
An object of class \code{BayesRates_model_ar1}.
}
\description{
Specify an AR1 (autoregression) model
for a time effect. The prior assumes that the
time effect will converge to a log-run average
specified by the user. The user can specify
whether this convergence is fast or slow.
}
\section{Mathematical details}{


The model for the time effect, on the log scale, is

\deqn{\beta_t = \phi \beta_{t-1} + (1 - \phi) \alpha + \epsilon_t}

where
\itemize{
\item \eqn{\beta_t} is the time effect
\item \eqn{\phi} is the autocorrelation coefficient
\item \eqn{\alpha} is the long-run average
\item \eqn{\epsilon} is a normally-distributed error term,
with mean \eqn{0} and standard deviation \eqn{\sigma}.
}

Parameter \eqn{\phi} has prior

\deqn{(\phi - m_0)/(m_1 - m_0) \sim \text{Beta}(2, 2)}

with values for \eqn{m_0} and \eqn{m_1} being supplied
by the user.

Parameter \eqn{\alpha} has prior

\deqn{\alpha \sim \text{N}(m, s^2)}

with values for \eqn{m} and \eqn{s} being supplied
by the user.

Standard deviation \eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, a^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for \eqn{a} is supplied by the user.
}

\examples{
AR1()
AR1(mean = 2)
}
\seealso{
\itemize{
\item \code{\link[=LocalTrend]{LocalTrend()}} creates a "local trend" model
for time effects
\item \code{\link[=RW2]{RW2()}} creates a random walk model for
age effects
\item \code{\link[=Spline]{Spline()}} creates a spline model for
age effects
}

The default bounds for the autocorrelation coefficient
are based on the bounds in \code{forecast::est()}.
}
