% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_setup_sparse.R
\name{basis_setup_sparse}
\alias{basis_setup_sparse}
\title{set up spline basis for sparse data}
\usage{
basis_setup_sparse(sfpca_data, nknots, orth = TRUE, delta = 1/10000)
}
\arguments{
\item{sfpca_data}{longitudinal data generated from prepare_data() function (list)}

\item{nknots:}{user-defined number of knots}

\item{orth:}{default setting for orth should be TRUE (after discussed with Wes on 02/13/2019)}

\item{delta:}{set to be 1/10000 to avoid rounding error(need eddit)}
}
\value{
A list containing the spline basis of sparse data
}
\description{
set up spline basis for sparse data
}
