% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesRates-package.R
\docType{package}
\name{BayesRates-package}
\alias{BayesRates}
\alias{BayesRates-package}
\title{Bayesian smoothing over and and time}
\description{
Bayesian smoothing of demographic rates over age and time. Calculations are done internally using TMB.
}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

Authors:
\itemize{
  \item Junni Zhang \email{junnizhang@163.com}
}

}
\keyword{internal}
