% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cn_divorces}
\alias{cn_divorces}
\title{Divorces in China}
\format{
A data frame with the following variables:
\itemize{
\item \code{age}: Age, in single-year age groups
\item \code{sex}: \code{"Female"} or \code{"Male"}
\item \code{time}: Calendar year
\item \code{nevent}: Numbers of divorces during year.
}
}
\source{
Tabulation from China Family Panel Studies
(CFPS) survey data.
}
\usage{
cn_divorces
}
\description{
Counts of divorces among respondents
in China Family Panel Studies survey, 1980-2018
}
\seealso{
\itemize{
\item \code{\link{cn_population}} population at risk
\item \code{\link{nz_divorces}} New Zealand data
}
}
\keyword{datasets}
