% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cn_population}
\alias{cn_population}
\title{Adult population in China}
\format{
A data frame with the following variables:
\itemize{
\item \code{age}: Age, in single-year age groups.
\item \code{sex}: \code{"Female"} or \code{"Male"}.
\item \code{time}: Calendar year.
\item \code{py}: Person-years lived.
}
}
\source{
Tabulation from China Family Panel Studies
(CFPS) survey data.
}
\usage{
cn_population
}
\description{
Person-years lived by respondents in
China Family Panel Studies survey, 1980-2018.
}
\seealso{
\itemize{
\item \code{\link{cn_divorces}} numbers of divorces
\item \code{\link{nz_population}} New Zealand data
}
}
\keyword{datasets}
