% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{components.BayesRates_results}
\alias{components.BayesRates_results}
\title{Extract components from a results object}
\usage{
\method{components}{BayesRates_results}(object, what, quantiles = TRUE, ...)
}
\arguments{
\item{object}{A results object.}

\item{what}{Component to be extracted.}

\item{...}{Not currently used.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Extract components from an object
of class \code{"BayesRates_results"}.
}
\details{
The components that can be extracted are:
\itemize{
\item \code{"rates"} Rates disaggregated by variables specified
in \code{agevar}, \code{timevar}, and \code{byvar} arguments.
\item \code{"age_effect"} Age main effect.
\item \code{"time_effect"} Time main effect.
\item \code{"age_hyper"} Hyper-parametrs for age main effect.
\item \code{"time_hyper"} Hyper-parametrs for time main effect.
}
}
\examples{
results <- smooth.agetime(nevent_df = nz_divorces,
                          py_df = nz_population,
                          spec_age = RW2(),
                          spec_time = AR1())
components(results, what = "age_effect")
components(results, what = "time_effect")
}
\seealso{
\code{\link[=augment]{augment()}}
}
