% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ECAM}
\alias{ECAM}
\title{Description(***)}
\format{
A data frame with 875 rows and 7 variables:
\describe{
\item{ID_unique}{the unqiue subject ID in study}
\item{Time}{(***)}
\item{response}{***}
\item{abx}{***}
\item{delivery}{the delivery mode of each subject}
\item{diet}{***}
\item{sex}{the gender of subject}
...
}
}
\source{
\url{***}
}
\usage{
data(ECAM)
}
\description{
A dataset containing 43 subjects(***)
}
\keyword{datasets}
