% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfpca_rda.R
\name{effectSize_rda}
\alias{effectSize_rda}
\title{Function to perform rda analysis on fpc score from model}
\usage{
effectSize_rda(output, model, variables)
}
\arguments{
\item{output:}{The output list from output_results() function}

\item{model:}{A set of variable names insterested in original data}

\item{variables:}{A list of interested variables for rda analysis}
}
\value{
A list of results from sfpca model and function
}
\description{
Function to perform rda analysis on fpc score from model
}
