% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invariants.R
\name{invariants}
\alias{invariants}
\title{generate sfpca models with different parameters}
\usage{
invariants(data, unique_subject_id, time_name, var_name = NULL)
}
\arguments{
\item{data:}{A data frame with subject id, time and variables}

\item{unique_subject_id:}{A column name corresponding to unique
subject id in the data (string)}

\item{time_name:}{A column name corresponding to time variable}

\item{var_name:}{A vector of interested variable name, default is null}
}
\value{
A vector of column names in data that do not change with
time variable
}
\description{
generate sfpca models with different parameters
}
\examples{
data("ECAM")
invariants(ECAM, 'studyid', 'month_of_life')
}
