% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-constructors.R
\name{LocalTrend}
\alias{LocalTrend}
\title{Specify a damped local trend model for a time effect}
\usage{
LocalTrend(
  scale_effect = 1,
  scale_level = 1,
  scale_trend = 1,
  coef_min = 0.8,
  coef_max = 0.98
)
}
\arguments{
\item{scale_effect}{Scale
for error for time effect. Defaults to 1.}

\item{scale_level}{Scale
for error for level term. Defaults to 1.}

\item{scale_trend}{Scale for prior
for error for trend term. Defaults to 1.}

\item{coef_min, coef_max}{Lower and upper limits
for the autocorrelation coefficient.
Defaults are 0.80 and 0.98.}
}
\value{
An object of class \code{BayesRates_model_localtrend}.
}
\description{
Specify a local trend model for a time effect.
The model assumes that the time effect follows
a random walk with drift, where the drift
itself can also change over time.
}
\section{Mathematical details}{


The model for the time effect, on the log scale, is

\deqn{\beta_t = \alpha_{t-1} + \epsilon_t^{(\beta)}}
\deqn{\alpha_t = \alpha_{t-1} + \delta_{t-1} + \epsilon_t^{(\alpha)}}
\deqn{\delta_t = \phi \delta_{t-1} + \epsilon_t^{(\delta)}}

where
\itemize{
\item \eqn{\beta_t} is the time effect
\item \eqn{\alpha} is a 'level' term
\item \eqn{\delta} is a 'trend' term
\item \eqn{\phi} is an autocorrelation coefficient
\item \eqn{\epsilon^{(\beta)}} is normally-distributed error term
with mean 0 and standard deviation \eqn{\sigma^{(\beta)}}
\item \eqn{\epsilon^{(\alpha)}} is normally-distributed error term
with mean 0 and standard deviation \eqn{\sigma^{(\alpha)}}
\item \eqn{\epsilon^{(\delta)}} is normally-distributed error term
with mean 0 and standard deviation \eqn{\sigma^{(\delta)}}
}

Parameter \eqn{\phi} has prior

\deqn{(\phi - m_0)/(m_1 - m_0) \sim \text{Beta}(2, 2)}

with values for \eqn{m_0} and \eqn{m_1} being supplied
by the user.

Parameter \eqn{\sigma^{(\beta)}} has prior

\deqn{\sigma^{(\beta)} \sim \text{N}^+(0, a_{\beta}^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for \eqn{a_{\beta}} is supplied by the user.

Parameter \eqn{\sigma^{(\alpha)}} has prior

\deqn{\sigma^{(\alpha)} \sim \text{N}^+(0, a_{\alpha}^2)}

and parameter \eqn{\sigma^{(\delta)}} has prior

\deqn{\sigma^{(\delta)} \sim \text{N}^+(0, a_{\delta}^2)}
}

\examples{
LocalTrend()
LocalTrend(scale_trend = 0.1)
}
\seealso{
\itemize{
\item \code{\link[=AR1]{AR1()}} creates a autoregression model
for time effects
\item \code{\link[=RW2]{RW2()}} creates a random walk model for
age effects
\item \code{\link[=Spline]{Spline()}} creates a spline model for
age effects
}
}
