% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{n_draw<-}
\alias{n_draw<-}
\title{Set the number of draws from the posterior distribution}
\usage{
n_draw(x) <- value
}
\arguments{
\item{x}{An object of class \code{"BayesRates_results"},
generated by a call to \code{\link[=smooth_age]{smooth_age()}} or \code{\link[=smooth_agetime]{smooth_agetime()}}.}

\item{value}{Number of posterior draws. A non-negative
integer.}
}
\value{
A modified version of \code{x}.
}
\description{
Specify the number of draws from the posterior
distribution that are generated by a results object from
\code{\link[=smooth_age]{smooth_age()}} or \code{\link[=smooth_agetime]{smooth_agetime()}}.
}
\examples{
res <- smooth_age(nevent_df = nz_divorces_2020,
                  py_df = nz_population_2020)
res
n_draw(res) <- 20
res

}
