% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{n_draw}
\alias{n_draw}
\alias{n_draw<-}
\alias{n_draw<-.BayesRates_results}
\title{Get and set the number of draws from the posterior distribution}
\usage{
n_draw(x)

n_draw(x) <- value

\method{n_draw}{BayesRates_results}(x) <- value
}
\arguments{
\item{x}{An object of holding results, eg
an object generated by a call to
\code{\link[=smooth_age]{smooth_age()}} or \code{\link[=smooth_agetime]{smooth_agetime()}}.}

\item{value}{Number of posterior draws. A non-negative
integer.}
}
\value{
A modified version of \code{x}.
}
\description{
Query or specify the number of draws from the posterior
distribution that are generated by a results object.
}
\examples{
res <- smooth_age(nevent_df = nz_divorces_2020,
                  py_df = nz_population_2020,
                  age_width_df = nz_age_width_df,
                  age_min = 15)
n_draw(res)
n_draw(res) <- 20
n_draw(res)

}
\seealso{
\itemize{
\item \code{\link[=smooth_age]{smooth_age()}} smooths rates over age.
\item \code{\link[=smooth_agetime]{smooth_agetime()}} smooths rates over age and time.
\item \code{\link[=augment.BayesRates_results]{augment()}}
combines data and estimates for rates.
\item \code{\link[=components.BayesRates_results]{components()}}
extracts rates, age effects,
time effects, and hyper-parameters.
\item \code{\link[=total_rate]{total_rate()}} calculates a summary indicator
}
}
