% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nz_age_width_df}
\alias{nz_age_width_df}
\title{Widths of age groups used with divorce data
for New Zealand}
\format{
A data frame with the following variables:
\itemize{
\item \code{age}: Age, in 5-year age groups, \code{"15-19"} to \code{"65+"}.
\item \code{width}: Width of the age group.
}
}
\usage{
nz_age_width_df
}
\description{
Widths of age groups used by
\code{\link{nz_divorces}}, \code{\link{nz_population}},
\code{\link{nz_divorces_2020}},
and \code{\link{nz_population_2020}}.
}
\seealso{
Data frames like \code{nz_age_width_df} are
used by functions \code{\link[=smooth_age]{smooth_age()}} and \code{\link[=smooth_agetime]{smooth_agetime()}}
when dealing with multi-year age groups.
}
\keyword{datasets}
