% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nz_divorces_2020}
\alias{nz_divorces_2020}
\title{Divorces in New Zealand in 2020}
\format{
A data frame with the following variables:
\itemize{
\item \code{age}: Age, in 5-year age groups, \code{"15-19"} to \code{"65+"}.
\item \code{sex}: \code{"Female"} or \code{"Male"}.
\item \code{nevent}: Numbers of divorces during year.
}
}
\source{
Derived from data in table "Age at divorces by
sex (marriages and civil unions) (Annual-Dec)"
in the online database Infoshare
on the Statistics New Zealand website.
Data downloaded on 22 March 2023.
}
\usage{
nz_divorces_2020
}
\description{
Counts of divorces by age and sex,
in New Zealand, 2020.
}
\seealso{
\itemize{
\item \code{\link{nz_divorces}} data for 1992-2021
\item \code{\link{nz_population_2020}} population at risk
}
}
\keyword{datasets}
