% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nz_population}
\alias{nz_population}
\title{Adult population of New Zealand}
\format{
A data frame with 660 rows and the
following variables:
\itemize{
\item \code{age}: Age, in 5-year age groups, 15-19 to 65+
\item \code{sex}: \code{"Female"} or \code{"Male"}
\item \code{time}: Calendar year
\item \code{py}: Person-years lived
}
}
\source{
Derived from data in table
"Estimated Resident Population by Age and Sex (1991+)
(Annual-Dec)" in the online
database Infoshare
on the Statistics New Zealand website.
Data downloaded on 26 March 2023.
}
\usage{
nz_population
}
\description{
Person-years lived by population aged 15+ by age, sex,
and calendar year, in New Zealand, 1992-2021
}
\keyword{datasets}
