% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nz_population_2020}
\alias{nz_population_2020}
\title{Adult population of New Zealand}
\format{
A data frame with the following variables:
\itemize{
\item \code{age}: Age, in 5-year age groups, 15-19 to 65+
\item \code{sex}: \code{"Female"} or \code{"Male"}
\item \code{py}: Person-years lived
}
}
\source{
Derived from data in table
"Estimated Resident Population by Age and Sex (1991+)
(Annual-Dec)" in the online
database Infoshare
on the Statistics New Zealand website.
Data downloaded on 26 March 2023.
}
\usage{
nz_population_2020
}
\description{
Person-years lived by population aged 15+
by age and sex, in New Zealand, 2020
}
\seealso{
\itemize{
\item \code{\link{nz_population}} data 1992-2021
\item \code{\link{nz_divorces_2020}} numbers of divorces
}
}
\keyword{datasets}
