% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.R
\name{optimal}
\alias{optimal}
\title{A function to compare sfpca model restuls and return the optimal one}
\usage{
optimal(model_list)
}
\arguments{
\item{model_list:}{A list of sfpca models with different parameters}
}
\value{
One sfpca model with the optimal result
}
\description{
A function to compare sfpca model restuls and return the optimal one
}
