% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_results.R
\name{output_results}
\alias{output_results}
\title{generate sfpca models with different parameters}
\usage{
output_results(data, variables, basis, rotation)
}
\arguments{
\item{data:}{The prepared data from prepare_data() function}

\item{variables:}{A set of variable names insterested in original data}

\item{basis:}{The list of basis results from basis_setup_sparse()}

\item{rotation:}{The list of rotation results from post_hoc_rotation()}
}
\value{
A list of results from sfpca model and function
}
\description{
generate sfpca models with different parameters
}
