% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostic.R
\name{plot_density_overlay}
\alias{plot_density_overlay}
\title{A function to draw density overlay plot}
\usage{
plot_density_overlay(sfpca_data, model)
}
\arguments{
\item{sfpca_data:}{The prepared data from prepare_data() function (list)}

\item{model:}{The optimal sfpca model}
}
\description{
A function to draw density overlay plot
}
