% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostic.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{A function to draw diganostic plot}
\usage{
plot_diagnostic(sfpca_data, sfpca_model)
}
\arguments{
\item{sfpca_data:}{The prepared data from prepare_data() function (list)}

\item{sfpca_model:}{The optimal sfpca model}
}
\description{
A function to draw diganostic plot
}
