% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_boxplot.R
\name{plot_fpc_boxplot}
\alias{plot_fpc_boxplot}
\title{plot boxplot of model fpc scores by group
Suggestion: take out missing values beforehand}
\usage{
plot_fpc_boxplot(
  output,
  pc_idx,
  group_name,
  testing_type = "parametric",
  pairwise_testing = FALSE,
  adjust_meth = "bonferroni"
)
}
\arguments{
\item{group_name}{One column name of interested group variables in data}

\item{output:}{The model output list from output_results() function}
}
\description{
plot boxplot of model fpc scores by group
Suggestion: take out missing values beforehand
}
