% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_group_mean.R
\name{plot_fpc_group_mean}
\alias{plot_fpc_group_mean}
\title{plot group mean of fpc scores}
\usage{
plot_fpc_group_mean(output, group_name)
}
\arguments{
\item{group_name}{One column name of interested group variables in data}

\item{output:}{The model output list from output_results() function}
}
\description{
plot group mean of fpc scores
}
