% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostic.R
\name{plot_k_diagnostic}
\alias{plot_k_diagnostic}
\title{A function to draw k diganostic plot}
\usage{
plot_k_diagnostic(da_list, model)
}
\arguments{
\item{data:}{The prepared data from prepare_data() function (list)}

\item{model:}{The optimal sfpca model}
}
\description{
A function to draw k diganostic plot
}
