% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_curve.R
\name{plot_mean_curve}
\alias{plot_mean_curve}
\title{A function to draw diganostic plot}
\usage{
plot_mean_curve(
  sfpca_data,
  basis,
  output,
  id_name,
  time_name,
  response_name,
  variables,
  ymin = NULL
)
}
\arguments{
\item{sfpca_data:}{The prepared data from prepare_data() function (list)}

\item{sfpca_model:}{A list of sfpca models with different parameters}
}
\description{
A function to draw diganostic plot
}
