% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs_predict.R
\name{plot_obs_predict}
\alias{plot_obs_predict}
\title{compare predicted curves (mean with 95\% credible intervals) to observed curves}
\usage{
plot_obs_predict(
  sfpca_data,
  optimal_model,
  data,
  time_name,
  response_name,
  unique_subject_id,
  subject_selected,
  ymin,
  ymax,
  x_lab,
  y_lab,
  title = "observed vs. predicted"
)
}
\arguments{
\item{sfpca_data:}{The prepared data list from prepare_data() function}

\item{optimal_model:}{optimal sfpca model}

\item{data:}{original dataset}

\item{time_name:}{time variable from original dataset}

\item{response_name:}{response variable from origina dataset}

\item{unique_subject_id:}{unique subject id identifier}

\item{subject_selected:}{interested subject for plotting}

\item{ymin:}{minimum value of y axis range}

\item{ymax:}{maximum value of y axis range}

\item{x_lab:}{label for x axis}

\item{y_lab:}{label for y axis}
}
\description{
compare predicted curves (mean with 95\% credible intervals) to observed curves
}
