% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostic.R
\name{plot_posterior_diagnostic}
\alias{plot_posterior_diagnostic}
\title{A function to draw density overlay plot}
\usage{
plot_posterior_diagnostic(sfpca_data, model, x_lab = NULL, y_lab = NULL)
}
\arguments{
\item{sfpca_data:}{The prepared data from prepare_data() function (list)}

\item{model:}{The optimal sfpca model}

\item{x_lab:}{Manually set x axis title}

\item{y_lab:}{Manually set y axis title}
}
\value{
A list with the plot
}
\description{
A function to draw density overlay plot
}
