% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qqPlot.R
\name{plot_residual}
\alias{plot_residual}
\title{A function to generate qqplot of the response variable}
\usage{
plot_residual(output)
}
\arguments{
\item{output:}{the outputlist with dataframe 'df'}
}
\description{
A function to generate qqplot of the response variable
}
