% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_rotation.R
\name{post_hoc_rotation}
\alias{post_hoc_rotation}
\title{perform post hoc rotation}
\usage{
post_hoc_rotation(da_list, model)
}
\arguments{
\item{model}{The optimal sfpca model}

\item{da_list:}{The prepared data from prepare_data() function}
}
\value{
A list after post hoc rotation (***)
}
\description{
perform post hoc rotation
}
