% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-constructors.R
\name{RW2}
\alias{RW2}
\title{Specify a random walk model for the age effect}
\usage{
RW2(scale = 1)
}
\arguments{
\item{scale}{Scale for error term.
Defaults to 1.}
}
\value{
An object of class \code{BayesRates_spec_rw2}.
}
\description{
Specify a second-order random walk model
for the age effect. A second-order
random walk tends to be smoother than an
ordinary (first-order) random walk,
while still being flexible enough to
capture distinctive features of the age profile
being modelled.
}
\section{Mathematical details}{


The model for the age effect, on the log scale, is

\deqn{(\beta_a - \beta_{a-1}) - (\beta_{a-1} - \beta_{a-2}) \sim \text{N}(0, \sigma^2)}

Parameter \eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, a^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for \eqn{a} is supplied by the user.
}

\examples{
RW2()
RW2(scale = 0.1)
}
\seealso{
\itemize{
\item \code{\link[=Spline]{Spline()}} is an alternative approach to
modelling age effects, based on splines.
\item \code{\link[=TimeFixed]{TimeFixed()}} is used to model time trends where the
age pattern is fixed over time.
\item \code{\link[=TimeVarying]{TimeVarying()}} is used to model time trends where the
age pattern varies over time.
}
}
