% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfpca_fit.R
\name{sfpca_fit}
\alias{sfpca_fit}
\title{generate sfpca models with different parameters}
\usage{
sfpca_fit(sfpca_data, Nsamples, Nchains, PC_max, D_max)
}
\arguments{
\item{PC_max:}{Number of the maximum principal component}

\item{D_max:}{Number of the maximum knot}

\item{Nsamples:}{Number of objects sampling from rstan}

\item{Nchains:}{Number of Markov chain using in rstan model}

\item{smod:}{Stan model constructed from source file (rstan)}

\item{sfpca_data:}{The prepared data list from prepare_data() function}
}
\value{
A list of sfpca classes with difference pc and knot numbers
}
\description{
generate sfpca models with different parameters
}
