% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfpca_plot_diagnostic.R
\name{sfpca_plot_diagnostic}
\alias{sfpca_plot_diagnostic}
\title{A function to draw diganostic plot}
\usage{
sfpca_plot_diagnostic(sfpca_data, sfpca_model)
}
\arguments{
\item{sfpca_data:}{The prepared data from prepare_data() function (list)}

\item{sfpca_model:}{The optimal sfpca model}
}
\description{
A function to draw diganostic plot
}
