% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqPlot.R
\name{sfpca_plot_residual}
\alias{sfpca_plot_residual}
\title{A function to generate qqplot of the response variable}
\usage{
sfpca_plot_residual(output)
}
\arguments{
\item{output:}{the outputlist with dataframe 'df'}
}
\description{
A function to generate qqplot of the response variable
}
