% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqPlot.R
\name{sfpca_qqplot}
\alias{sfpca_qqplot}
\title{A function to generate qqplot of the response variable}
\usage{
sfpca_qqplot(data, response_name)
}
\arguments{
\item{data:}{A dataframe with the response column to plot}

\item{response:}{A column name of the interested response in data}
}
\description{
A function to generate qqplot of the response variable
}
