% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\name{sfpca_rda}
\alias{sfpca_rda}
\title{Function to perform rda analysis on fpc score from model}
\usage{
sfpca_rda(output, model, variables = NULL)
}
\arguments{
\item{output:}{The prepared data from prepare_data() function}

\item{model:}{A set of variable names insterested in original data}

\item{basis:}{The list of basis results from basis_setup_sparse()}
}
\value{
A list of results from sfpca model and function
}
\description{
Function to perform rda analysis on fpc score from model
}
