% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.agetime.R
\name{smooth.agetime}
\alias{smooth.agetime}
\title{Smooth rates over age and time}
\usage{
smooth.agetime(
  nevent_df,
  py_df,
  agevar = "age",
  timevar = "time",
  byvar = character(),
  model_age = RW2(),
  model_time = Spline(),
  n_draw = 1000L
)
}
\arguments{
\item{nevent_df}{A data frame with a column called \code{"nevent"}
containing the number of events, plus additional classification
variables.}

\item{py_df}{A data frame with a column called \code{"py"}
containing exposure, plus additional
classification variables.}

\item{agevar}{The name of the age variable in \code{nevent_df} and \code{py_df}.
Defaults to \code{"age"}.}

\item{timevar}{The name of the time variable in \code{nevent_df} and \code{py_df}.
Defaults to \code{"time"}.}

\item{byvar}{The names of classification variables in \code{nevent_df}
and \code{py_df}, other than age and time (eg \code{"sex"} or \code{"region"}).
Optional.}

\item{model_age}{The prior model for the age dimension.
Current choices: \code{"Spline"}, \code{"RW2"} (second-order random walk).}

\item{model_time}{The prior model for the time dimension. Current choices:
\code{"AR1"}, \code{"LocalTrend"}.}

\item{n_draw}{Number of draws from posterior distribution
to use in output. Defaults to 1000.}
}
\value{
An object of class \code{"BayesRates_smooth_agetime"}.
}
\description{
Smooth rates over age and time
}
\seealso{
\code{\link[=rates_age]{rates_age()}}, \code{\link[=rates_total]{rates_total()}}
}
