% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_agetime.R
\name{smooth_agetime}
\alias{smooth_agetime}
\title{Smooth rates over age and time}
\usage{
smooth_agetime(
  nevent_df,
  py_df,
  timevar = "time",
  agevar = "age",
  byvar = character(),
  model_time = c("AR1", "LocalTrend"),
  model_age = c("Spline", "RW2")
)
}
\arguments{
\item{nevent_df}{A data frame with a column called \code{"nevent"}
containing the number of events, plus additional classification
variables.}

\item{py_df}{A data frame with a column called \code{"py"}
containing exposure, plus additional
classification variables.}

\item{timevar}{The name of the time variable in \code{nevent_df} and \code{py_df}.
Defaults to \code{"time"}.}

\item{agevar}{The name of the age variable in \code{nevent_df} and \code{py_df}.
Defaults to \code{"age"}.}

\item{byvar}{The names of classification variables in \code{nevent_df}
and \code{py_df}, other than age and time (eg \code{"sex"} or \code{"region"}).
Optional.}

\item{model_time}{The prior model for the time dimension. Current choices:
\code{"AR1"}, \code{"LocalTrend"}.}

\item{model_age:}{The prior model for the age dimension.
Current choices: \code{"Spline"}, \code{"RW2"} (second-order random walk).}
}
\value{
An object of class \code{"BayesTime_smooth_agetime"}.
}
\description{
Smooth rates over age and time
}
\seealso{
\code{\link[=rates_age]{rates_age()}}, \code{\link[=rates_total]{rates_total()}}
}
