% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-constructors.R
\name{Spline}
\alias{Spline}
\title{Specify a P-spline model for an age effect}
\usage{
Spline(df = 5L, scale = 1)
}
\arguments{
\item{df}{Degrees of freedom for spline.
Defaults to 5.}

\item{scale}{Scale for error term.
Defaults to 1.}
}
\value{
An object of class \code{BayesRates_model_spline}.
}
\description{
Specify a P-spline (penalised spline)
model for an age effect.
A P-spline model is flexible, but
favours profiles that are relatively smooth.
}
\section{Mathematical details}{


The model for the age effect, on the log scale, is

\deqn{\beta = X \alpha}

where
\itemize{
\item \eqn{\beta} is a vector of age effects
\item \eqn{X} is a matrix holding the basis functions
for the spline, with \code{df} columns
\item \eqn{\alpha} is a vector of coefficients,
with \code{df} elements
}

The elements of \eqn{\alpha} are assumed to follow
a second order random walk,

\deqn{(\alpha_a - \alpha_{a-1}) - (\alpha_{a-1} - \alpha_{a-2}) \sim \text{N}(0, \sigma^2)}

Parameter \eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, a^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for \eqn{a} is supplied by the user.
}

\examples{
Spline()
Spline(df = 10)
}
\seealso{
\itemize{
\item \code{\link[=AR1]{AR1()}} creates a autoregression model
for time effects
\item \code{\link[=LocalTrend]{LocalTrend()}} creates a local trend model
for time effects
\item \code{\link[=RW2]{RW2()}} creates a random walk model for
age effects
}

\code{Spline()} uses function \code{\link[splines:bs]{splines::bs()}} to
generate matrix \eqn{X}.
}
