% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\name{stan_fit}
\alias{stan_fit}
\title{generate sfpca models with different parameters}
\usage{
stan_fit(
  sfpca_data,
  Nsamples,
  Nchains,
  Ncores = NULL,
  PC_range,
  nknot_range,
  seed = NULL
)
}
\arguments{
\item{sfpca_data:}{The prepared data list from prepare_data() function}

\item{Nsamples:}{Number of objects sampling from rstan}

\item{Nchains:}{Number of Markov chain using in rstan model}

\item{Ncores:}{Number of cores using in rstan model}

\item{PC_range:}{A vector of pc number}

\item{nknot_range:}{A vector of knot number}
}
\value{
A list of sfpca classes with difference pc and knot numbers
}
\description{
generate sfpca models with different parameters
}
