% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-constructors.R
\name{TimeFixed}
\alias{TimeFixed}
\title{Specify time trends: age pattern fixed}
\usage{
TimeFixed(scale = 1)
}
\arguments{
\item{scale}{Scale for the standard deviation
of the error term. Default is 1.}
}
\value{
An object of class \code{"BayesRates_spec_timefixed"}.
}
\description{
Specify time trends where the shape of
the age profile is assumed to be constant
over time, even though the level
may change. The level is assumed to follow
a random walk.
}
\section{Mathematical details}{


In the model underlying \code{TimeFixed()},

\deqn{\eta_{at} = \alpha_a + \beta_t}

where
\itemize{
\item \eqn{\eta_{at}} denotes (log) rates
for age group \eqn{a} during period \eqn{t},
\item \eqn{\alpha_a} is an age effect,
\item \eqn{\beta_t} is a time effect,
}

and the time effect is modelled as

\deqn{\beta_t = \beta_{t-1} + \epsilon_t}

where \eqn{\epsilon} is a normally-distributed error term
with mean \eqn{0} and standard deviation \eqn{\sigma}.

Standard deviation \eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, a^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for scale parameter \eqn{a}
is supplied by the user.
}

\examples{
TimeFixed()
TimeFixed(scale = 2)
}
\seealso{
\itemize{
\item \code{\link[=TimeVarying]{TimeVarying()}} is an alternative approach to modelling
time trends. \code{TimeVarying()} allows age profiles to
change over time.
\item \code{\link[=RW2]{RW2()}} models age effects using a random walk.
\item \code{\link[=Spline]{Spline()}} models age effects using splines.
}
}
