% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{total_rate}
\alias{total_rate}
\title{Summarise rates by adding up over the life time}
\usage{
total_rate(x, interval = 0.95)
}
\arguments{
\item{x}{An object of class \code{"BayesRates_results"},
constructed via a call to \code{\link[=smooth_age]{smooth_age()}} or
\code{\link[=smooth_agetime]{smooth_agetime()}}.}

\item{interval}{Width of credible intervals.
A number between 0 and 1. Default is \code{0.95}.}
}
\value{
A tibble with the following columns:
\itemize{
\item the time variable specified by \code{timevar},
if \code{x} was created by a call to \code{\link[=smooth_agetime]{smooth_agetime()}}
\item further classifying variables, as specified by
\code{byvar} in the original call to \code{\link[=smooth_age]{smooth_age()}} or
\code{\link[=smooth_agetime]{smooth_agetime()}}
\item \code{.fitted}. Point estimates from the model.
(Posterior medians.)
\item \code{.lower}, \code{.upper}. Lower and upper limits
of the credible interval specified by \code{interval}.
\item \code{.probability}. A list column containing all draws
from the posterior distribution.
\item \code{.observed}. Total rates constructed from
direct estimates, ie events divided by population,
with no smoothing.
}
}
\description{
Construct a summary measure for age-specific rates
by summing up the rates, weighted by the width of each
age group. The calculations are equivalent to those
used to construct the total fertility rate (TFR).
}
\examples{
res <- smooth_agetime(nevent_df = cn_divorces,
                      py_df = cn_population,
                      spec_time = TimeFixed())
total_rate(res)
}
\seealso{
\itemize{
\item \code{\link[=smooth_age]{smooth_age()}} smooths rates over age
\item \code{\link[=smooth_agetime]{smooth_agetime()}} smooths rates over age and time
\item \code{\link[=augment]{augment()}} combines data and estimates for rates
\item \code{\link[=components]{components()}} extracts rates, age effects,
time effects, and hyper-parameters
\item \code{\link[=n_draw<-]{n_draw<-()}} sets the default number
of draws from the posterior distribution
}
}
