% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{augment.BayesRates_results}
\alias{augment.BayesRates_results}
\title{Data and values from a results object}
\usage{
\method{augment}{BayesRates_results}(x, ...)
}
\arguments{
\item{x}{A \code{"BayesRates_results"} object created
by calling function \code{\link[=smooth.agetime]{smooth.agetime()}}.}

\item{...}{Not currently used.}
}
\value{
A \link[tibble:tibble-package]{tibble},
constructed by merging the original \code{nevent_df} and
\code{py_df} arguments to \code{\link[=smooth.agetime]{smooth.agetime()}},
plus four new columns:
\itemize{
\item \code{.fitted} Point estimates (posterior medians) of rates.
\item \code{.lower}, \code{.upper} Lower and upper bounds of
95\% credible intervals for \code{.fitted}.
\item \code{.observed} Direct estimates of the rate.
}
}
\description{
Data and values from a results object
}
\examples{
results <- smooth.agetime(nevent_df = nz_divorces,
                          py_df = nz_population,
                          model_age = RW2(),
                          model_time = AR1())
augment(results)
}
