% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesRates-package.R
\docType{package}
\name{BayesRates-package}
\alias{BayesRates}
\alias{BayesRates-package}
\title{Smooth demographic rates using Bayesian methods}
\description{
Use Bayesian methods to smooth demographic rates
over age and (optionally) over time. Calculations
are done internally using package TMB.
}
\details{
The key functions are
\itemize{
\item \code{\link[=smooth_age]{smooth_age()}} for smoothing over age only
\item \code{\link[=smooth_agetime]{smooth_agetime()}} for simultaneously smoothing
over age and time
}
}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

Authors:
\itemize{
  \item Junni Zhang \email{junnizhang@163.com}
}

}
\keyword{internal}
