% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{components.BayesRates_results}
\alias{components.BayesRates_results}
\title{Extract components from a results object}
\usage{
\method{components}{BayesRates_results}(object, what = NULL, interval = 0.95, ...)
}
\arguments{
\item{object}{A results object.}

\item{what}{Components to be extracted.
If no value supplied, all available components
are extracted.}

\item{interval}{Width of credible intervals.
A number between 0 and 1. Default is \code{0.95}.}

\item{...}{Not currently used.}
}
\value{
A named list of tibbles, or, if \code{what}
is a single value, a single tibble.
Each tibble contains
\itemize{
\item classification variables, if present
\item \code{.fitted} Point estimates (posterior medians).
\item \code{.lower}, \code{.upper} Lower and upper bounds of
credible intervals.
\item \code{.probability}. A list column with all draws
from the posterior distribution.
}

The age variable comes in versions: the original
version and one with a \code{.mid} suffix. The \code{.mid}
version is useful for plotting.
}
\description{
Extract components from an object
created by a call to \code{\link[=smooth_age]{smooth_age()}}
or \code{\link[=smooth_agetime]{smooth_agetime()}}.
}
\details{
If the object was created by a call to
\code{\link[=smooth_age]{smooth_age()}}, then the components that can be extracted are:
\itemize{
\item \code{"rates"} Age-specific rates.
\item \code{"intercept"} Model intercept.
\item \code{"age_effect"} Age main effect.
\item \code{"age_hyper"} Hyper-parameters for age main effect.
}

If the object was created by a call to \code{\link[=smooth_agetime]{smooth_agetime()}},
then the components that can be extracted are
\itemize{
\item \code{"rates"} Age-time-specific rates
\item \code{"intercept"} Model intercept.
\item \code{"age_effect"} Age main effect.
\item \code{"time_effect"} Time main effect or age-time interaction.
\item \code{"age_hyper"} Hyper-parametres for age main effect.
\item \code{"time_hyper"} Hyper-parameters for time main effect or
age-time interaction.
}
}
\examples{
## run model
results <- smooth_age(nevent_df = nz_divorces_2020,
                      py_df = nz_population_2020,
                      age_width_df = nz_age_width_df,
                      age_min = 15)

## extract all components
components(results)

## extract age effect
components(results, what = "age_effect")
}
\seealso{
\itemize{
\item \code{\link[=smooth_age]{smooth_age()}} smooths rates over age.
\item \code{\link[=smooth_agetime]{smooth_agetime()}} smooths rates over age and time.
\item \code{\link[=augment.BayesRates_results]{augment()}}
combines data and estimates for rates.
\item \code{\link[=n_draw<-]{n_draw()}} sets the default number
of draws from the posterior distribution.
\item \code{\link[=total_rate]{total_rate()}} calculates a summary indicator
}
}
