% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nz_divorces}
\alias{nz_divorces}
\title{Divorces in New Zealand}
\format{
A data frame with 660 rows and the
following variables:
\itemize{
\item \code{age}: Age, in 5-year age groups, 15-19 to 65+
\item \code{sex}: \code{"Female"} or \code{"Male"}
\item \code{time}: Calendar year
\item \code{nevent}: Numbers of divorces during year.
}
}
\source{
Derived from data in table "Age at divorces by
sex (marriages and civil unions) (Annual-Dec)"
in the online database Infoshare
on the Statistics New Zealand website.
Data downloaded on 22 March 2023.
}
\usage{
nz_divorces
}
\description{
Counts of divorces by age, sex,
and calendar year, in New Zealand, 1992-2021
}
\keyword{datasets}
