% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_divorces.R
\docType{data}
\name{nz_divorces}
\alias{nz_divorces}
\title{Divorces and population in New Zealand}
\format{
A data frame with 660 rows and the
following variables:
\itemize{
\item \code{age}: Age, in 5-year age groups, 15-19 to 65+
\item \code{sex}: \code{"Female"} or \code{"Male"}
\item \code{time}: Calendar year
\item \code{divorces}: Numbers of divorces during year.
\item \code{population}: Estimated average population during year.
}
}
\source{
Derived from data in tables "Age at divorces by
sex (marriages and civil unions) (Annual-Dec)"
and "Estimated Resident Population by Age and Sex (1991+)
(Annual-Dec)" in the online
database Infoshare
on the Statistics New Zealand website.
Data downloaded on 22 March and 26 March 2023.
}
\usage{
nz_divorces
}
\description{
Counts of divorces and of population by age, sex,
and calendar year, in New Zealand, 1992-2021
}
\keyword{datasets}
