% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_results.R
\name{output_results}
\alias{output_results}
\title{generate sfpca models with different parameters}
\usage{
output_results(sfpca_data, model)
}
\arguments{
\item{sfpca_data:}{The prepared data from prepare_data() function}

\item{model:}{The optimal model from optimal() function}
}
\value{
A list of results from model
}
\description{
generate sfpca models with different parameters
}
