% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_results.R
\name{output_results}
\alias{output_results}
\title{generate sfpca models with different parameters}
\usage{
output_results(da_list, model)
}
\arguments{
\item{da_list:}{The prepared data from prepare_data() function}

\item{model:}{The optimal model from optimal() function}

\item{variables:}{A set of variable names insterested in original data}
}
\value{
A list of results from model
}
\description{
generate sfpca models with different parameters
}
