% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_boxplot.R
\name{plot_fpc_boxplot}
\alias{plot_fpc_boxplot}
\title{plot boxplot of model fpc scores by group
Suggestion: take out missing values beforehand}
\usage{
plot_fpc_boxplot(
  output,
  pc_idx,
  group_name,
  x_lab = NULL,
  y_lab = NULL,
  p_title = NULL,
  testing_type = c("parametric", "non-parametric"),
  pairwise_testing = FALSE,
  global_testing = FALSE,
  group_order = NULL
)
}
\arguments{
\item{output}{The model output list from output_results() function}

\item{pc_idx}{The pc index number}

\item{group_name}{One column name of interested group variables in data}

\item{x_lab}{Manually set x axis title}

\item{y_lab}{Manually set y axis title}

\item{p_title}{Manually set plot title}

\item{testing_type}{The option to set (non-)parametric test}

\item{pairwise_testing}{The option to set pairwise test}

\item{global_testing}{The option to set global test}

\item{group_order}{The string vector for boxplot order}
}
\description{
plot boxplot of model fpc scores by group
Suggestion: take out missing values beforehand
}
