% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_curve.R
\name{plot_fpc_curve}
\alias{plot_fpc_curve}
\title{A function to plot fpc curves}
\usage{
plot_fpc_curve(data, response_name, basis, output, rotation)
}
\arguments{
\item{data}{A data frame for longitudinal analysis}

\item{response_name}{The name of response variable}

\item{basis}{The model basis results list from basis_setup_sparse() function}

\item{output}{The model output list from output_results() function}
}
\description{
A function to plot fpc curves
}
