% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_curve.R
\name{plot_fpc_curve}
\alias{plot_fpc_curve}
\title{A function to plot fpc curves}
\usage{
plot_fpc_curve(
  output,
  pc_idx,
  original = FALSE,
  x_lab = NULL,
  y_lab = NULL,
  ymin = NULL,
  ymax = NULL,
  x_tick = NULL
)
}
\arguments{
\item{output}{The model output list from output_results() function}

\item{pc_idx}{The indices of pc to plot}

\item{x_lab}{Manually set x axis title}

\item{y_lab}{Manually set y axis title}

\item{ymin}{The minimum of y lab}

\item{ymax}{The maximum of y lab}

\item{x_tick}{Manually set x tick}
}
\description{
A function to plot fpc curves
}
