% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_curve.R
\name{plot_fpc_curve}
\alias{plot_fpc_curve}
\title{A function to plot fpc curves}
\usage{
plot_fpc_curve(output, original = FALSE, ymin = NULL, ymax = NULL)
}
\arguments{
\item{output}{The model output list from output_results() function}

\item{original}{The option to plot with original or transformed
time and response value}
}
\description{
A function to plot fpc curves
}
