% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_curve.R
\name{plot_fpc_curve}
\alias{plot_fpc_curve}
\title{A function to plot fpc curves}
\usage{
plot_fpc_curve(output, pc_idx, ymin = NULL, ymax = NULL)
}
\arguments{
\item{output}{The model output list from output_results() function}

\item{pc_idx}{The indices of pc to plot}

\item{ymin}{The minimum of y lab}

\item{ymax}{The maximum of y lab}
}
\description{
A function to plot fpc curves
}
