% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_group_mean.R
\name{plot_fpc_group_mean}
\alias{plot_fpc_group_mean}
\title{plot group mean of fpc scores}
\usage{
plot_fpc_group_mean(
  output,
  pc_idx,
  original = FALSE,
  group_name,
  x_lab = NULL,
  y_lab = NULL,
  ymin = NULL,
  ymax = NULL
)
}
\arguments{
\item{pc_idx}{The pc index to plot with}

\item{original}{The option to plot with original or transformed}

\item{group_name}{One column name of interested group variables in data}

\item{ymin}{The minimum of y lab}

\item{ymax}{The maximum of y lab}

\item{output:}{The model output list from output_results() function}

\item{x_lab:}{Manually set x axis title}

\item{y_lab:}{Manually set y axis title}
}
\value{
A list with the plot and the data used for plot
}
\description{
plot group mean of fpc scores
}
