% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_group_mean.R
\name{plot_fpc_group_mean}
\alias{plot_fpc_group_mean}
\title{plot group mean of fpc scores}
\usage{
plot_fpc_group_mean(
  output,
  pc_idx,
  original = FALSE,
  group_name,
  ymin = NULL,
  ymax = NULL
)
}
\arguments{
\item{pc_idx}{The pc index to plot with}

\item{original}{The option to plot with original or transformed}

\item{group_name}{One column name of interested group variables in data}

\item{output:}{The model output list from output_results() function}
}
\description{
plot group mean of fpc scores
}
