% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc_on_mean_curve.R
\name{plot_fpc_on_mean_curve}
\alias{plot_fpc_on_mean_curve}
\title{A function to plot fpc on mean curves}
\usage{
plot_fpc_on_mean_curve(output, pc_idx, original = FALSE)
}
\arguments{
\item{output}{The model output list from output_results() function}

\item{pc_idx}{The pc index to plot with}

\item{original}{The option to plot with original or transformed
time and response value}
}
\description{
A function to plot fpc on mean curves
}
