% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group.R
\name{plot_group}
\alias{plot_group}
\title{A function to generate plots of a group of variables
Suggestion: take out missing values beforehand}
\usage{
plot_group(data, time_name, response_name, unique_subject_id, variable_name)
}
\arguments{
\item{unique_subject_id}{the column name corresponding to unique subject id in the data (string)}

\item{variable_name}{One column name of interested variable from dataset}

\item{data:}{A dataframe with the response column to plot}

\item{time_name:}{the column name corresponding to the time variable in the data (string)}

\item{response_name:}{the column name of the intersted response variable (string)}
}
\value{
A list with the plot and the data used for plot
}
\description{
A function to generate plots of a group of variables
Suggestion: take out missing values beforehand
}
