% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_curve.R
\name{plot_mean_curve}
\alias{plot_mean_curve}
\title{A function to plot mean curces}
\usage{
plot_mean_curve(output, ymin = NULL, ymax = NULL)
}
\arguments{
\item{output:}{The model output from output_results() function}

\item{ymin:}{Manually set minimum of y lab}

\item{ymax:}{Manually set maximum of y lab}
}
\description{
A function to plot mean curces
}
