% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_curve.R
\name{plot_mean_curve}
\alias{plot_mean_curve}
\title{A function to plot mean curces}
\usage{
plot_mean_curve(
  output,
  original = FALSE,
  xlab = NULL,
  ylab = NULL,
  ymin = NULL,
  ymax = NULL
)
}
\arguments{
\item{output:}{The model output from output_results() function}

\item{original:}{The option to plot with original or transformed}

\item{ymin:}{Manually set minimum of y lab}

\item{ymax:}{Manually set maximum of y lab}
}
\description{
A function to plot mean curces
}
