% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_rotation.R
\name{post_hoc_rotation}
\alias{post_hoc_rotation}
\title{perform post hoc rotation}
\usage{
post_hoc_rotation(sfpca_data, model)
}
\arguments{
\item{model}{The picked sfpca model}

\item{data:}{The prepared data from prepare_data() function}
}
\value{
A list after post hoc rotation (need change)
}
\description{
perform post hoc rotation
}
