% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-constructors.R
\name{RW2}
\alias{RW2}
\title{Specify a random walk model for an age effect}
\usage{
RW2(scale = 1)
}
\arguments{
\item{scale}{Scale for error term.
Defaults to 1.}
}
\value{
An object of class \code{BayesRates_model_rw2}.
}
\description{
Specify a second-order random walk model
for an age effect. In a second-order random
walk the change in increments, rather than the
increments themselves, are random. A second-order
random walk tends to be smoother than an
ordinary (first-order) random walk.
}
\section{Mathematical details}{


The model for the age effect, on the log scale, is

\deqn{(\beta_a - \beta_{a-1}) - (\beta_{a-1} - \beta_{a-2}) \sim \text{N}(0, \sigma^2)}

Parameter \eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, a^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for \eqn{a} is supplied by the user.
}

\examples{
RW2()
RW2(scale = 0.1)
}
\seealso{
\itemize{
\item \code{\link[=AR1]{AR1()}} creates a autoregression model
for time effects
\item \code{\link[=LocalTrend]{LocalTrend()}} creates a local trend model
for time effects
\item \code{\link[=Spline]{Spline()}} creates a spline model for
age effects
}
}
