% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\name{sfpcaClass}
\alias{sfpcaClass}
\title{S3 class sfpca model with number of pc, knot, sampling data from rstan,
log_liklihood, and leave-one-out cross-validation (LOO) information criterion}
\usage{
sfpcaClass(
  Nsamples = NULL,
  Nchains = NULL,
  pc = NULL,
  knot = NULL,
  sa = NULL,
  log_lik = NULL,
  looic = NULL,
  basis = NULL
)
}
\description{
S3 class sfpca model with number of pc, knot, sampling data from rstan,
log_liklihood, and leave-one-out cross-validation (LOO) information criterion
}
