% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_agetime.R
\name{smooth_agetime}
\alias{smooth_agetime}
\title{Smooth rates over age and time}
\usage{
smooth_agetime(
  nevent_df,
  py_df,
  agevar = "age",
  timevar = "time",
  byvar = character(),
  age_width_df = NULL,
  age_min = NULL,
  spec_age = Spline(),
  spec_time = TimeVarying()
)
}
\arguments{
\item{nevent_df}{A data frame with a column
called \code{"nevent"} containing the number of events,
plus additional classification variables.}

\item{py_df}{A data frame with a column called \code{"py"}
containing exposure, plus additional
classification variables.}

\item{agevar}{The name of the age variable in
\code{nevent_df} and \code{py_df}.
Defaults to \code{"age"}.}

\item{timevar}{The name of the time variable in
\code{nevent_df} and \code{py_df}.
Defaults to \code{"time"}.}

\item{byvar}{The names of classification variables in \code{nevent_df}
and \code{py_df}, other than age and time (eg \code{"sex"} or \code{"region"}).
Optional.}

\item{age_width_df}{A data frame with columns \code{"age"}
and \code{"width"}, giving the widths of each age group.
Required if the age group labels are not integers.}

\item{age_min}{The lower limit of the youngest age group.
Required if the age group labels are not integers.}

\item{spec_age}{The prior model for the age effect.
Current choices: \code{\link[=Spline]{Spline()}} and \code{\link[=RW2]{RW2()}}. Defaults to
\code{Spline()}.}

\item{spec_time}{The prior model for the time effect.
Current choices: \code{\link[=TimeFixed]{TimeFixed()}} and \code{\link[=TimeVarying]{TimeVarying()}}.
Defaults to \code{TimeVarying()}.}
}
\value{
An object of class \code{"BayesRates_results"}.
}
\description{
Smooth rates over age and time
}
