% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-constructors.R
\name{Spline}
\alias{Spline}
\title{Specify a P-spline model for an age effect}
\usage{
Spline(df = NULL, scale = 1)
}
\arguments{
\item{df}{Degrees of freedom for spline.
The default is 0.7 times the number of age groups,
or 4, whichever is higher.}

\item{scale}{Scale for error term.
The default is 1.}
}
\value{
An object of class \code{"BayesRates_spec_spline"}.
}
\description{
Specify a P-spline (penalised spline)
model for an age effect.
A P-spline is flexible, but
favours profiles that are relatively smooth.
}
\section{Mathematical details}{


The model for the age effect, on the log scale, is

\deqn{\beta = X \gamma}

where
\itemize{
\item \eqn{\beta} is a vector of age effects,
\item \eqn{X} is a matrix holding the basis functions
for the spline, with \code{df} columns, and
\item \eqn{\gamma} is a vector of coefficients,
with \code{df} elements.
}

The elements of \eqn{\gamma} are assumed to follow
a second order random walk,

\deqn{(\gamma_a - \gamma_{a-1}) - (\gamma_{a-1} - \gamma_{a-2}) \sim \text{N}(0, \sigma^2)}

Parameter \eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, a^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for \eqn{a} is supplied by the user.
}

\examples{
Spline()
Spline(df = 10)
}
\seealso{
\itemize{
\item \code{\link[=RW2]{RW2()}} is an alternative approach to
modelling age effects, based on a random walk.
\item \code{\link[=TimeFixed]{TimeFixed()}} specifies time trends where the
age pattern is fixed over time.
\item \code{\link[=TimeVarying]{TimeVarying()}} specifies time trends where the
age pattern varies over time
}

\code{Spline()} uses function \code{\link[splines:bs]{splines::bs()}} internally.
}
