% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\name{stan_fit}
\alias{stan_fit}
\title{generate sfpca models with different parameters}
\usage{
stan_fit(da_list, Nsamples, Nchains, PC_max, D_max)
}
\arguments{
\item{da_list:}{The prepared data list from prepare_data() function}

\item{Nsamples:}{Number of objects sampling from rstan}

\item{Nchains:}{Number of Markov chain using in rstan model}

\item{PC_max:}{Number of the maximum principal component}

\item{D_max:}{Number of the maximum knot}
}
\value{
A list of sfpca classes with difference pc and knot numbers
}
\description{
generate sfpca models with different parameters
}
