% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-constructors.R
\name{TimeFixed}
\alias{TimeFixed}
\title{Specify time trends: age pattern fixed}
\usage{
TimeFixed(scale = 1)
}
\arguments{
\item{scale}{Scale for the standard deviation
of the error term. Defaults to 1.}
}
\value{
An object of class \code{"BayesRates_spec_timefixed"}.
}
\description{
Specify time trends where the age pattern is
assumed to maintain the same shape over time,
and the time effect follows a random walk.
}
\section{Mathematical details}{


When the age pattern is fixed, the
model for log rates is

\deqn{\eta_{at} = \alpha_a + \beta_t}

where
\itemize{
\item \eqn{\eta_{at}} denotes the rates (on a log scale)
for age group \eqn{a} during period \eqn{t}
\item \eqn{\alpha_a} is an age effect
\item \eqn{\beta_t} is a time effect
}

and the time effect is modelled as

\deqn{\beta_t = \beta_{t-1} + \epsilon_t}

where \eqn{\epsilon} is a normally-distributed error term,
with mean \eqn{0} and standard deviation \eqn{\sigma}.

Standard deviation \eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, a^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for scale parameter \eqn{a}
is supplied by the user.
}

\examples{
TimeFixed()
TimeFixed(scale = 2)
}
\seealso{
\itemize{
\item \code{\link[=TimeVarying]{TimeVarying()}} specifies time trends
where the age pattern varies
\item \code{\link[=RW2]{RW2()}} creates a random walk specification for
age effects
\item \code{\link[=Spline]{Spline()}} creates a spline specification for
age effects
}
}
