% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-constructors.R
\name{TimeVarying}
\alias{TimeVarying}
\title{Specify time trends: age pattern changes}
\usage{
TimeVarying(scale = 1)
}
\arguments{
\item{scale}{Scale for the standard deviation
of the error term. Default is 1.}
}
\value{
An object of class \code{"BayesRates_spec_timevarying"}.
}
\description{
Specify time trends where the age pattern
is allowed to vary smoothly over time.
}
\section{Mathematical details}{


In the model underlying \code{TimeVarying()},

\deqn{\eta_{at} = \alpha_a + \beta_{at}}

where
\itemize{
\item \eqn{\eta_{at}} denotes (log) rates
for age group \eqn{a} during period \eqn{t},
\item \eqn{\alpha_a} is an age effect,
\item \eqn{\beta_{at}} is an age-time interaction,
}

and the age-time interaction is modelled as

\deqn{\beta_t = \beta_{t-1} + \epsilon_t}

where
\itemize{
\item \eqn{\beta_t} is a vector of age-specific terms, and
\item \eqn{\epsilon_t} is a vector drawn from a multivariate
normal with mean 0 and variance \eqn{V}
}

Element \eqn{v_{ij}} of \eqn{V} has the form

\deqn{v_{ij} = \rho^{|i-j|}\sigma^2}

Correlation coefficient \eqn{\rho} has a Beta(2,2) prior.

Standard deviation \eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, a^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for scale parameter \eqn{a}
is supplied by the user.
}

\examples{
TimeVarying()
TimeVarying(scale = 0.1)
}
\seealso{
\itemize{
\item \code{\link[=TimeFixed]{TimeFixed()}} is an alternative approach to modelling
time trends. \code{TimeFixed()} assumes that age profiles are
constant over time.
\item \code{\link[=RW2]{RW2()}} models age effects using a random walk.
\item \code{\link[=Spline]{Spline()}} models age effects using splines.
}
}
